requirements.txt
requests
selenium
airtable-python-wrapper
bs4
lxml